refactor(#127): five files had outgrown the 600-line law
The CI size check was failing on five files. Each one is split by what its
parts are ABOUT, not at an arbitrary line, and nothing moves but position:
no test dropped, renamed or weakened, no production behaviour touched.

- application: the parameter-write cases (Number/Bool/Text/Option) leave
  local_dispatcher_tests.rs for ld_block_param_tests.rs, the same
  `ld_*_tests.rs` shape the earlier splits use. The root keeps the shared
  fixtures the whole family glob-imports, and its own ToggleBlockEnabled tests.
- adapter-gui: io_bindings.rs keeps the pure helpers, the Slint model
  projection and `wire`; what a gesture actually DOES (WireCtx + the callback
  installers for both window surfaces) moves to io_bindings_callbacks.rs.
  WireCtx and `create_binding` become pub(super) so io_bindings_tests can
  still build one without an AppWindow.
- adapter-gui: the ui_index_to_real_block_index cases move to
  lib_ui_index_tests.rs; the chain fixtures stay in lib_tests.rs, which
  lib_recent_tests already shares.
- engine: the output-side callback (process_output_f32 and _mixed) moves to
  runtime_output_process.rs, re-exported from runtime.rs so
  `engine::runtime::process_output_f32{,_mixed}` is unchanged. Whole items,
  same order, same bodies — nothing on the audio path is re-implemented.
- engine: #85's cross-rate oracle keeps its eight assertions; the rig, the
  two-clock driver and the residual measurement move to
  issue_85_other_rate_harness/, a plain module of the same test binary, so
  every case still runs where it did.

validate.sh: 5 errors -> 0 (18 pre-existing debt warnings unchanged).
cargo test --workspace: 4290 passed / 0 failed / 31 ignored, before and after.
cargo build --workspace: no warnings.
